已经完成了留言功能，但因无网络无法验证，明日再验证

+++ myblog/models/posts.js
@@ -1,5 +1,29 @@
 const Post = require('../lib/mongo').Post
 const marked = require('marked')
+const CommentModel = require('../models/comments')
+
+
+Post.plugin('addCommentsCount',{
+    afterFind:function(posts){
+        return Promise.all(posts.map(function(post){
+            return CommentModel.getCommentsCount(post._id).then(function(commentsCount){
+                post.commentsCount = commentsCount
+                return post
+            })
+        }))
+    },
+    afterFindOne:function (post){
+        if(post){
+            return CommentModel.getCommentsCount(post._id).then(function(count){
+                post.commentsCount = count
+                return post
+            })
+        }
+        return post
+    }
+})
+
+
 
 Post.plugin('contentToHtml',{
     afterFind:function(posts){
@@ -24,7 +48,7 @@ module.exports = {
     },
     //通过文章id获取文章
     getPostById:function getPostById(postId){
-        return Post.findOne({_id:postId}).populate({path:'author',model:'User'}).addCreatedAt().contentToHtml().exec()
+        return Post.findOne({_id:postId}).populate({path:'author',model:'User'}).addCreatedAt().addCommentsCount().contentToHtml().exec()
     },
     //按创建的时间降序获取所有用户文章或者某个特定用户的所有文章
     getPosts: function getPosts (author) {
@@ -37,6 +61,7 @@ module.exports = {
           .populate({ path: 'author', model: 'User' })
           .sort({ _id: -1 })
           .addCreatedAt()
+          .addCommentsCount()
           .contentToHtml()
           .exec()
       },
@@ -56,7 +81,12 @@ module.exports = {
         return Post.update({_id:postId},{$set:data}).exe()
     },
     //通过文章id删除一篇文章
-    delPostById:function delPostById(postId){
-        return Post.deleteOne({_id:postId}).exe()
+    delPostById:function delPostById(postId,author){
+        return Post.deleteOne({author:author,_id:postId}).exe().then(function(res){
+            //文章删除后，再删除该文章下的所有留言
+            if(res.result.ok && res.result.n > 0 ){
+                return CommentModel.delCommentByPostId(postId)
+            }
+        })
     }
 }
